工作报告更新progress

+++ com.bizvisionsoft.service/src/com/bizvisionsoft/service/model/WorkInfo.java
@@ -593,7 +593,7 @@ public class WorkInfo {
 	@ReadValue
 	@WriteValue
 	@Persistence
-	private Float progress;
+	private Double progress;
 	////////////////////////////////////////////////////////////////////////////////////////////////////
 
 	@ReadValue
